Moved node helpers into `utils/nodes.ts`
